+++ src/main/java/com/softserveinc/orphanagemenu/validator/warehouse/WarehouseItemForm.java
@@ -3,13 +3,9 @@ package com.softserveinc.orphanagemenu.validator.warehouse;
 public class WarehouseItemForm {
 	private Long id;
 	private String itemName;
-	private Float quantity;
+	private Double quantity;
 	private String dimension;
 	
-
-	/* (non-Javadoc)
-	 * @see java.lang.Object#toString()
-	 */
 	@Override
 	public String toString() {
 		return "WarehouseItemForm [id=" + id + ", itemName=" + itemName
@@ -29,10 +25,10 @@ public class WarehouseItemForm {
 	public void setItemName(String itemName) {
 		this.itemName = itemName;
 	}
-	public Float getQuantity() {
+	public Double getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(Float quantity) {
+	public void setQuantity(Double quantity) {
 		this.quantity = quantity;
 	}
 	public String getDimension() {
